requirements.txt
openai
pandas
numpy
pyplotlib
pillow
jsonschema